all: new v2-unstable version (#104)
We use the macaroon.v2-unstable package, allowing us to
have binary caveat ids and macaroon ids.

We change some of the API accordingly, removing some elements
that are unnecessary, such as the first party location argument
to the AcquireDischarge method.

Note that this changes the format of root keys stored by the mgostorage
package. There is some code to partially deal with that by falling back
to fetching from the old format, but a subsequent PR will make that more
comprehensive.

This also adds the ability for a third party discharger to be able to
see the keys used to decrypt the third party caveat, which will
enable some proposed use cases.