bakery, httpbakery: support v2 macaroons and caveat ids.
This has a bunch of changes which together
mean that the bakery supports binary-encoded
macaroons and caveat ids. This causes cookie
sizes to drop to around half their previous size (~1600 bytes
to ~800 bytes). That size will continue to drop in subsequent PRs.

Note that there is still one failing test which will require
an update to httprequest for a fix. Rather than delaying
this PR further, we intend to land this as an interim branch
and fix the discharge fallback behaviour in the next PR.

Rename PublicKeyLocator to ThirdPartyLocator and make
it convey information on the version of the third party too.

Remove PublicKeyRing as it's complex and has marginal use.

Use version 2 macaroons when possible, while staying backwardly
compatible with old clients and dischargers.

Remove functions added only because we couldn't change
the signature of existing functions (e.g. CheckAnyM)

Add version arguments to methods where the version is
important (e.g. bakery.Service.NewMacaroon).

Add version information to local third party caveat
so we know what kind of caveat encoding to use.

Changed bakerytest.NewDischarger to be more generally
useful, and added a helper to make it straightforward to
get parsed caveats as before.

Use httprequest to implement the HTTP discharge handler,
and implement the new endpoints needed to implement
the ThirdPartyLocator interface.