bakery/checkers: refactor to use contexts and namespaces
We change the way that first party caveat checking works
so that instead of requiring a closure or object to be created
containing the request context (thus requiring a checker
to be created afresh for each request), we use an context.Context
to hold information used for verification by checkers.

This means we can treat checkers as much more static objects.
The idea is that any application will probably have only one
instance of a checker object with all the checks supported
by first party caveats in that application.

We also implement namespaces (as discussed in the Hague sprint).
A checkers.Caveat now contains a Namespace field containing a URI
of the namespace to use. This is resolved by a Namespace instance
when actually adding the caveat to the macaroon.

The delete the idservice example because it's a pain to fix and
we want to rewrite it to use the upcoming auth model anyway.