httpbakery: deprecate DoWithBody
DoWithBody is awkward and only exists because we
couldn't think of a way around the fact that http.NewRequest
hides the original type when used on seekable types
such as *bytes.Reader.

We work around that here by observing that we can actually
get the underlying reader out of the result of NopCloser,
meaning that we can just make the standard Do method work
as nature intended.