bakery: implement checker type
This is the start of a replacement for the existing Service type. The
Service type does not pull its weight - it is only a thin wrapper around
functionality already exposed as functions/ It does not provide a useful
abstraction that makes it straightforward to build secure services with
macaroons. It also presupposes that there is a single authorization space
for macaroons - any macaroon can potentially authorize any operation,
limited only by first party caveats.

The new Checker type is a strict superset of the checking functionality
provided by Service. Every macaroon is associated one or more
*operations*, where an operation specifies (by name) an entity and an
action to take on that entity. The current Service behaviour is the same
as always creating macaroons with the LoginOp operation.

We bless our existing conventions for leveraging external identity
providers to authorize requests. A request may be authorized by virtue
of declared identity, but it also may be authorized by macaroons that
are associated with the required operations.

With the new Checker in place, it becomes possible to safely create
delegatable macaroons ("capabilities") that authorize a set of operations,
and it also becomes possible to arbitrarily combine these capabilities.

The Checker type does not itself create macaroons - this is left up to
the user of the Checker. There is another type that will be introduced
in a subsequent PR that will provide this functionality, at which point
the Service type can be removed.