bakery: add Oven and Bakery types
The Oven is the type responsible for creating macaroons and verifying
their signatures. Responsibility for checking caveats and operations
is left to the Checker. The Oven should be compatible with old macaroon
ids too - if an old id is found, it is assumed to be associated with the
LoginOp operation, as prior to this all macaroons have been essentially
treated as login macaroons.

We use protocol buffers to encode the new macaroon identifiers,
which gives us a nicely compact format that we don't have to invent
ourselves - note that macaroon ids are considered opaque
so only the service generating them will need to be able to read them,
so we don't need to worry about python and javascript support for
protocol buffers.

Although technically both Oven and Checker are independently useful,
many applications will want to combine them, so we make a new transparent
Bakery type that puts them together for convenience. Some might argue
that Bakery could embed both types rather than just as fields - I don't
mind either way.

I've marked the examples to be ignored for now so that they don't clutter
this already huge PR.  To be good examples, they will be better off
rewritten to take advantage of the new API rather than just patched up.

We also rename the word "storage" to "store" throughout as there
is no good reason for the distinction - it's only confusing.