bakery: pass third party caveats externally
We define a new bakery.Macaroon type that holds external
third party caveat information and also the namespace
to use when adding first party caveats.

This is backwardly compatible with the old data format. When
a bakery.Macaroon containing an old macaroon is marshaled,
it uses the old format; when it's unmarshaled, it understands
the old formats. This should make it possible to
be fully backward compatible with old clients.