bakery/checkers: simplify declarations
As implied by the proposal document, we make declaration caveats simpler
and hopefully a little less error prone.

In the new scheme, instead of being able to declare an arbitrary
set of attribute/value pairs, each third party will in general use
only a single caveat to declare its information.

This should be less vulnerable to security flaws due to
attributes being added, because all attributes should be declared
within the same caveat.

For backward compatibility, we still support a limited form
of the old scheme, sufficient to declare a single attribute.
This should be sufficient for our existing identity manager
which always uses "declared username" and declares no other
attributes.

The need-declared mechanism is due to change too - we
still support the old scheme, but will implement the new scheme
in a subsequent PR.
 @rogpeppe