bakery: define new encoding format for third party caveats
Instead of a string, we use a protobuf encoding
that includes the need-declared information as
a field and the actual condition body as a binary
byte slice.

We change the Caveat struct so that the third party
condition body is specified separately from the first
party condition and need-declared caveats are
also specified separately.