httpbakery: rework interaction-required protocol
We make the interaction-required error return all the
supported interaction methods directly in the response
rather than requiring the client to fetch the visit-URL
with "Accept: application/json".

All interaction methods now complete by producing
a "discharge token" which can then be presented to
the discharge endpoint to acquire the discharge macaroon.
It is the responsibility of the server to check that the
discharge token is acceptable for the provided third
party caveat.

There is no longer a requirement that all methods fit
the "visit/wait" protocol, which means that the
protocol is now capable of supporting a redirect-based
flow.

We split the bakerytest functionality up into two parts.
One is the Discharger functionality which many external
packages use; the other (Rendezvous) makes it easier
to test visit/wait-style interaction-required protocols.
The actual HTTP endpoints are left up to individual
clients to implement, as they vary a lot depending
on what is being tested.