httpbakery: update NewDischargeRequiredError
The current design stems from the fact that we needed
to keep the API compatible. Now that we have bakery.Macaroon,
we can tell the version from the macaroon that's being returned,
and we can also provide more useful defaults for the cookie
path and its name (the current default of creating a separate cookie
for every macaroon is not such a great idea).

This is an API-breaking change.