bakery: factor out identity checks
Identity checking via third party caveats and declaration caveats
is fragile and doesn't really fit as part of the core bakery
abstraction. This PR factors it out into another package
(bakery/identchecker) that provides an API almost identical
to the current bakery API, while leaving the underlying
operation checker functionality in the core bakery package.

In the process, it has become clear that the AllowAny primitive
is problematic so it's been removed.

Some API changes made in this PR:

- declaration caveats are treated as structural caveats, and
operate on the current macaroons being checked instead
of a map inferred beforehand. checkers.ContextWithMacaroons
is used to add the relevant context.

- support for allow and deny caveats is removed from
checkers, as these interacted unpleasantly with the operation
checker (and they didn't actually check full operations anyway,
so were misleading).

- LegacyMacaroonOp is added to bakery.OvenParams so that
it can know which operation to associate with legacy macaroons