bakery: remove multi-op support; rename MacaroonOpStore
With the advent of OpsAuthorizer, if there's a need to store
large numbers of operations in a macaroon, that's possible
outside of the Oven by allowing a "multi-" op to authorize
its component operations. This functionality is not used
by any existing service anyway, so delete it in the interests
of simplicity.

Because this functionality was the only reason that
Oven.NewMacaroon required an expiry time, we can
now remove that argument from the somwhat argument-overburdened
NewMacaroon method, but we'll do that in the next PR.

We also rename MacaroonOpStore to MacaroonVerifier
because MacaroonOpStore is a confusing name and
MacaroonVerifier better describes its role.
by using the OpsAuthorizer to authorize the