httpbakery/agent: use POST not GET for agent login
When the client doesn't use the shortcut agent login
method (because it's not setting an agent-login cookie
for every request), the protocol is somewhat different.
We use a POST, not a GET, and the field names have changed.

Unfortunately this is challenging to write a decent unit
test for, as the unit test inevitably assumes the protocol
that we're testing against.

To QA, add an agent user to https://api.staging.jujucharms.com/identity
with:

	user-admin put user-admin put-agent --agent-file $HOME/.agent/testagent.agent --idm-url https://api.staging.jujucharms.com/identity $agentname

Update github.com/rogpeppe/bhttp to use this PR as a dependency.

	bhttp --agent $agentfile https://api.staging.jujucharms.com/identity/v1/whoami