httpbakery: be more careful about cookie expiry times
When there is clock skew between server and client, we
can receive a discharged macaroon that is apparently expired,
even though it's valid on the server.

We work around that issue by using a session cookie
(which will not be stored permanently) when the expiry
time is short.

We also use a long expiry time for the cookie when the
macaroon has no expiry time, rather than using a session
cookie as we were doing before.