use configurable logging
This means we can use an externally defined logger
if needed. There are a few places where this was awkward
because there was no place to put the Logger, but
those log messages don't seem important enough
to justify a backwardly incompatible API or separate
entry points, so they've been removed.

Rethinking the actual log messages printed is left
for another PR.