feat(rootkeys): add new constructor for root key store
Previously, the only constructor we had for bakery.RootkeyStore took in
a Backing, but if that backing also implements ContextBacking, we prefer
that backing.

Changes being made in Juju mean tha we naturally wish to implement
ContextBacking. Our db methods require a context. However, to make use
of the existing NewStore constructor, we would need to implement stub
methods without contexts.

We do not wish to expose these.

Add a new constructor which directly takes in a ContextBacking as it's
backing. This means we will no longer need to implement stub
context-less methods.